requirement.txt
fastapi
uvicorn
ultralytics
python-multipart
pillow